Refactor to float64, add cost to nodes
Former-commit-id: c3dafeeba677f2c94fd5435c42208b0f19a03468

+++ internal/defrag/build_test.go
@@ -20,7 +20,7 @@ func Test_build(t *testing.T) {
 	c.Synthesis.Cost = map[int]config.SynthCost{
 		100000: {
 			Fixed:   true,
-			Dollars: float32(0),
+			Dollars: 0.0,
 		},
 	}
